+++ database/migrations/2021_01_13_230725_delete_column_type_from_orders_table.php
@@ -14,7 +14,9 @@ class DeleteColumnTypeFromOrdersTable extends Migration
     public function up()
     {
         Schema::table('orders', function (Blueprint $table) {
-            DB::statement("ALTER TABLE `orders` DROP COLUMN `type`");
+            if (Schema::hasColumn('orders', 'type')) {
+                DB::statement("ALTER TABLE `orders` DROP COLUMN `type`");
+            }
         });
     }
