+++ database/migrations/2022_04_02_085059_remove_type_column_from_webinar_chapters_table.php
@@ -15,7 +15,9 @@ class RemoveTypeColumnFromWebinarChaptersTable extends Migration
     public function up()
     {
         Schema::table('webinar_chapters', function (Blueprint $table) {
-            DB::statement("ALTER TABLE `webinar_chapters` DROP COLUMN `type`");
+            if (Schema::hasColumn('webinar_chapters', 'type')) {
+                DB::statement("ALTER TABLE `webinar_chapters` DROP COLUMN `type`");
+            }
         });
     }
 }
