+++ database/migrations/2024_10_19_154545_edit_forum_recommended_topics_table.php
@@ -13,18 +13,22 @@ return new class extends Migration {
     public function up()
     {
         Schema::table('forum_recommended_topics', function (Blueprint $table) {
-            \Illuminate\Support\Facades\DB::statement("ALTER TABLE `forum_recommended_topics` DROP COLUMN `title`");
+            if (Schema::hasColumn('forum_recommended_topics', 'title')) {
+                \Illuminate\Support\Facades\DB::statement("ALTER TABLE `forum_recommended_topics` DROP COLUMN `title`");
+            }
         });
 
-        Schema::create('forum_recommended_topic_translations', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedInteger('forum_recommended_topic_id');
-            $table->string('locale', 191)->index();
-            $table->string('title');
-            $table->text('subtitle');
+        if (!Schema::hasTable('forum_recommended_topic_translations')) {
+            Schema::create('forum_recommended_topic_translations', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->unsignedInteger('forum_recommended_topic_id');
+                $table->string('locale', 191)->index();
+                $table->string('title');
+                $table->text('subtitle');
 
-            $table->foreign('forum_recommended_topic_id', 'forum_recommended_topic_id_trans')->on('forum_recommended_topics')->references('id')->onDelete('cascade');
-        });
+                $table->foreign('forum_recommended_topic_id', 'forum_recommended_topic_id_trans')->on('forum_recommended_topics')->references('id')->onDelete('cascade');
+            });
+        }
     }
 
     /**
